@@ -11,21 +11,21 @@ here = Path(__file__).resolve().parent
 
 setup(
     name=name,
-    version='1.6',
+    version='1.7',
     description='Improved simple time scheduler based on standard sched',
     long_description=here.joinpath('README.md').read_text(),
     long_description_content_type='text/markdown',
-    url='https://github.com/bulletmark/{}'.format(name),
+    url=f'https://github.com/bulletmark/{name}',
     author='Mark Blakeney',
     author_email='mark.blakeney@bullet-systems.net',
     keywords='sched scheduler timer periodic cron crontab',
     license='GPLv3',
     py_modules=[module],
-    python_requires='>=3.4',
+    python_requires='>=3.6',
     classifiers=[
         'Programming Language :: Python :: 3',
     ],
     data_files=[
-        ('share/{}'.format(name), ['README.md']),
+        (f'share/{name}', ['README.md']),
     ],
 )
